chore: Separate SecretClass and TrustStore specs
