+++ rust/operator-binary/src/crd/secret_class/v1alpha2_impl.rs
