+++ rust/operator-binary/src/main.rs
@@ -76,7 +76,7 @@ async fn main() -> anyhow::Result<()> {
         stackable_operator::cli::Command::Crd => {
             SecretClass::merged_crd(crd::SecretClassVersion::V1Alpha2)?
                 .print_yaml_schema(built_info::PKG_VERSION, SerializeOptions::default())?;
-            TrustStore::merged_crd(crd::TrustStoreVersion::V1Alpha2)?
+            TrustStore::merged_crd(crd::TrustStoreVersion::V1Alpha1)?
                 .print_yaml_schema(built_info::PKG_VERSION, SerializeOptions::default())?;
         }
         stackable_operator::cli::Command::Run(SecretOperatorRun {
